+++ src/Keter/Proxy.hs
@@ -32,7 +32,6 @@ import Data.Version (showVersion)
 import GHC.Exts (fromString)
 import Keter.Common
 import Keter.Config
-import Keter.Config.Middleware
 import Keter.Context
 import Keter.HostManager qualified as HostMan
 import Keter.Rewrite qualified as Rewrite
@@ -76,15 +75,16 @@ import System.FilePath (FilePath)
 import WaiAppStatic.Listing (defaultListing)
 
 data ProxySettings = MkProxySettings
-  { -- | Mapping from virtual hostname to port number.
-    psHostLookup     :: ByteString -> IO (Maybe (ProxyAction, TLS.Credentials))
-  , psManager        :: !Manager
-  , psIpFromHeader   :: Bool
+  { -- | Mapping from virtual hostname to action, credentials
+    psHostLookup          :: ByteString -> IO (Maybe (ProxyAction, TLS.Credentials))
+  , psManager             :: !Manager
+  , psIpFromHeader        :: Bool
+  , psTrustForwardedFor   :: Wai.Request -> Bool
   , psConnectionTimeBound :: Int
-  , psHealthcheckPath :: !(Maybe ByteString)
-  , psUnknownHost    :: ByteString -> ByteString
-  , psMissingHost    :: ByteString
-  , psProxyException :: ByteString
+  , psHealthcheckPath     :: !(Maybe ByteString)
+  , psUnknownHost         :: ByteString -> ByteString
+  , psMissingHost         :: ByteString
+  , psProxyException      :: ByteString
   }
 
 makeSettings :: HostMan.HostManager -> KeterM KeterConfig ProxySettings
@@ -99,6 +99,7 @@ makeSettings hostman = do
     let psConnectionTimeBound = kconfigConnectionTimeBound * 1000
     let psIpFromHeader = kconfigIpFromHeader
     let psHealthcheckPath = encodeUtf8 <$> kconfigHealthcheckPath
+    let psTrustForwardedFor _req = True
     pure $ MkProxySettings{..}
     where
         psHostLookup = HostMan.lookupAction hostman . CI.mk
@@ -135,33 +136,35 @@ reverseProxy listener = do
 
 connectClientCertificates :: (ByteString -> IO (Maybe (ProxyAction, TLS.Credentials))) -> Bool -> WarpTLS.TLSSettings -> WarpTLS.TLSSettings
 connectClientCertificates hl session s =
-    let
-        newHooks = WarpTLS.tlsServerHooks s
-        -- todo: add nested lookup
-        newOnServerNameIndication (Just n) =
-             maybe mempty snd <$> hl (S8.pack n)
-        newOnServerNameIndication Nothing =
-             return mempty -- we could return default certificate here
-    in
-        s { WarpTLS.tlsServerHooks = newHooks{TLS.onServerNameIndication = newOnServerNameIndication}
-          , WarpTLS.tlsSessionManagerConfig = if session then Just TLSSession.defaultConfig else Nothing }
+  let
+    newHooks = WarpTLS.tlsServerHooks s
+    newOnSNI :: Maybe TLS.HostName -> IO TLS.Credentials
+    newOnSNI (Just n) = do
+      -- Return per-host certificate if available
+      mx <- hl (S8.pack n)
+      case mx of
+        Just ((_, _), creds) -> pure creds
+        Nothing              -> pure mempty
+    newOnSNI Nothing = pure mempty -- we could return default certificate here
+  in
+    s { WarpTLS.tlsServerHooks = newHooks { TLS.onServerNameIndication = newOnSNI }
+      , WarpTLS.tlsSessionManagerConfig = if session then Just TLSSession.defaultConfig else Nothing
+      }
 
 
-withClient :: Bool -- ^ is secure?
+withClient :: Bool -- ^ Is the outer listener secure (HTTPS)?
            -> KeterM ProxySettings Wai.Application
 withClient isSecure = do
     cfg@MkProxySettings{..} <- ask
-    let useHeader = psIpFromHeader
+    let honourHeader req = psIpFromHeader && psTrustForwardedFor req
     withRunInIO $ \rio ->
         pure $ waiProxyToSettings
            (error "First argument to waiProxyToSettings forced, even thought wpsGetDest provided")
            defaultWaiProxySettings
-            { wpsSetIpHeader =
-                if useHeader
-                    then SIHFromHeader
-                    else SIHFromSocket
-            ,  wpsGetDest = Just (getDest cfg)
-            ,  wpsOnExc = handleProxyException (\app e -> rio $ logException app e) psProxyException
+            { wpsSetIpHeader = SIHFromSocket  -- Default, will be overridden per-request in getDest
+            , wpsGetDest = Just (getDest cfg honourHeader)
+            , wpsOnExc =
+                 handleProxyException (\app e -> rio $ logException app e) psProxyException
             } psManager
   where
     logException :: Wai.Request -> SomeException -> KeterM ProxySettings ()
@@ -169,36 +172,36 @@ withClient isSecure = do
       "Got a proxy exception on request " <> show a <> " with exception "  <> show b
 
 
-    getDest :: ProxySettings -> Wai.Request -> IO (LocalWaiProxySettings, WaiProxyResponse)
+    getDest :: ProxySettings -> (Wai.Request -> Bool) -> Wai.Request -> IO (LocalWaiProxySettings, WaiProxyResponse)
     -- respond to healthckecks, regardless of Host header value and presence
-    getDest MkProxySettings{..} req | psHealthcheckPath == Just (Wai.rawPathInfo req)
+    getDest MkProxySettings{..} _ req
+      | psHealthcheckPath == Just (Wai.rawPathInfo req)
       = return (defaultLocalWaiProxySettings, WPRResponse healthcheckResponse)
     -- inspect Host header to determine which App to proxy to
-    getDest cfg@MkProxySettings{..} req =
+    getDest cfg@MkProxySettings{..} useHeaderPred req =
         case Wai.requestHeaderHost req of
             Nothing -> do
               return (defaultLocalWaiProxySettings, WPRResponse $ missingHostResponse psMissingHost)
-            Just host -> processHost cfg req host
+            Just host -> processHost cfg useHeaderPred req host
 
-    processHost :: ProxySettings -> Wai.Request -> S.ByteString -> IO (LocalWaiProxySettings, WaiProxyResponse)
-    processHost cfg@MkProxySettings{..} req host = do
+    processHost :: ProxySettings -> (Wai.Request -> Bool) -> Wai.Request -> S.ByteString -> IO (LocalWaiProxySettings, WaiProxyResponse)
+    processHost cfg@MkProxySettings{..} useHeaderPred req host = do
         -- Perform two levels of lookup. First: look up the entire host. If
         -- that fails, try stripping off any port number and try again.
-        mport <- liftIO $ do
-            mport1 <- psHostLookup host
-            case mport1 of
-                Just _ -> return mport1
-                Nothing -> do
-                    let host' = S.takeWhile (/= 58) host
-                    if host' == host
-                        then return Nothing
-                        else psHostLookup host'
-        case mport of
-            Nothing -> do -- we don't know the host that was asked for
-              return (defaultLocalWaiProxySettings, WPRResponse $ unknownHostResponse host (psUnknownHost host))
-            Just ((action, requiresSecure), _)
-                | requiresSecure && not isSecure -> performHttpsRedirect cfg host req
-                | otherwise -> performAction psManager isSecure psConnectionTimeBound req action
+        mEntry <- liftIO $ do
+          let tryHost h = psHostLookup h
+          m1 <- tryHost host
+          case m1 of
+            Just x -> pure (Just x)
+            Nothing -> do
+              let host' = S.takeWhile (/= 58) host
+              if host' == host then pure Nothing else tryHost host'
+        case mEntry of
+          Nothing -> do -- we don't know the host that was asked for
+            return (defaultLocalWaiProxySettings, WPRResponse $ unknownHostResponse host (psUnknownHost host))
+          Just ((action, requiresSecure), _cert)
+            | requiresSecure && not isSecure -> performHttpsRedirect cfg host req
+            | otherwise -> performAction psManager psProxyException isSecure useHeaderPred psConnectionTimeBound host req action
 
     performHttpsRedirect MkProxySettings{..} host =
         return . (addjustGlobalBound psConnectionTimeBound Nothing,) . WPRResponse . redirectApp config
@@ -226,31 +229,56 @@ addjustGlobalBound bound to = go `setLpsTimeBound` defaultLocalWaiProxySettings
            Just x | x > 0 -> Just x
            _              -> Nothing
 
+-- | Perform a stanza action, applying previously set up middlewares.
+performAction :: Manager
+              -> ByteString               -- ^ onExceptBody
+              -> Bool                     -- ^ isSecure
+              -> (Wai.Request -> Bool)    -- ^ should honour XFF?
+              -> Int                      -- ^ global bound (us)
+              -> ByteString               -- ^ host (unused now; kept for parity)
+              -> Wai.Request
+              -> ProxyActionRaw
+              -> IO (LocalWaiProxySettings, WaiProxyResponse)
+performAction psManager onExceptBody isSecure useHeaderPred globalBound _host req = \case
+  (PAPort port mw tbound) -> do
+    let innerApp :: Wai.Application
+        innerApp =
+          waiProxyToSettings
+            (error "inner default app forced, even though wpsGetDest is provided")
+            defaultWaiProxySettings
+              { wpsSetIpHeader =
+                  if useHeaderPred req then SIHFromHeader else SIHFromSocket
+              , wpsGetDest = Just $ \r -> do
+                  let protocol = if isSecure then "https" else "http"
+                      r' = r { Wai.requestHeaders =
+                                ("X-Forwarded-Proto", protocol) : Wai.requestHeaders r
+                             }
+                  pure (defaultLocalWaiProxySettings, WPRModifiedRequest r' (ProxyDest "127.0.0.1" port))
+              , wpsOnExc =
+                  handleProxyException (\_ _ -> pure ()) onExceptBody
+              } psManager
+    pure ( addjustGlobalBound globalBound tbound
+         , WPRApplication (mw innerApp)
+         )
 
-performAction :: Manager -> Bool -> Int -> Wai.Request -> ProxyActionRaw -> IO (LocalWaiProxySettings, WaiProxyResponse)
-performAction psManager isSecure globalBound req = \case
-  (PAPort port tbound) ->
-    return (addjustGlobalBound globalBound tbound, WPRModifiedRequest req' $ ProxyDest "127.0.0.1" port)
-      where
-        req' = req
-            { Wai.requestHeaders = ("X-Forwarded-Proto", protocol)
-                                : Wai.requestHeaders req
-            }
-        protocol
-            | isSecure = "https"
-            | otherwise = "http"
-  (PAStatic StaticFilesConfig {..}) ->
-    return (addjustGlobalBound globalBound sfconfigTimeout, WPRApplication $ processMiddleware sfconfigMiddleware $ staticApp (defaultFileServerSettings sfconfigRoot)
-        { ssListing =
-            if sfconfigListings
-                then Just defaultListing
-                else Nothing
-        })
-  (PARedirect config) -> return (addjustGlobalBound globalBound Nothing, WPRResponse $ redirectApp config req)
-  (PAReverseProxy config rpconfigMiddleware tbound) ->
-       return (addjustGlobalBound globalBound tbound, WPRApplication
-                $ processMiddleware rpconfigMiddleware
-                $ Rewrite.simpleReverseProxy psManager config
+  (PAStatic StaticFilesConfig {..} mw) -> do
+    let baseApp = staticApp (defaultFileServerSettings sfconfigRoot)
+          { ssListing =
+              if sfconfigListings
+                  then Just defaultListing
+                  else Nothing
+          }
+    return ( addjustGlobalBound globalBound sfconfigTimeout
+           , WPRApplication (mw baseApp)
+           )
+
+  (PARedirect config) ->
+    return (addjustGlobalBound globalBound Nothing, WPRResponse $ redirectApp config req)
+
+  (PAReverseProxy config mw tbound) -> do
+       let baseApp = Rewrite.simpleReverseProxy psManager config
+       return ( addjustGlobalBound globalBound tbound
+              , WPRApplication (mw baseApp)
               )
 
 redirectApp :: RedirectConfig -> Wai.Request -> Wai.Response