+++ ChangeLog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2.3.0
+
+- Add integration with keter-rate-limiting-plugin. [301](https://github.com/snoyberg/keter/issues/301)
+- Provided additional possibilities to apply middlewares per-app.
+
 ## 2.2.1
 - Fix unnecessary file reloading: https://github.com/snoyberg/keter/pull/320
 - bump tar